apps/transcribe-service/pyproject.toml
[project]
name = "window-transcribe-transcribe-service"
version = "0.1.0"
description = "ASR-only WhisperX transcription service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "loguru",
  "pydantic-settings",
  "uvicorn",
  "whisperx",
]

[project.scripts]
transcribe-service-api = "transcribe_service.server:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
